# openapi.yaml
openapi: 3.0.3
info:
  title: DPA Next API
  description: |
    DPA Next 智能研究助手 API
    
    提供项目生命周期管理、智能文档处理、认知分析等功能的RESTful API接口。
  version: 1.0.0
  contact:
    name: DPA开发团队
    email: support@dpa.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8001/api/v1
    description: 本地开发服务器
  - url: https://api.dpa.ai/v1
    description: 生产服务器
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      tags:
        - projects
      summary: 获取项目列表
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - name: project_type
          in: query
          schema:
            $ref: '#/components/schemas/ProjectType'
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
    
    post:
      tags:
        - projects
      summary: 创建项目
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: 项目创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{project_id}:
    get:
      tags:
        - projects
      summary: 获取项目详情
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      tags:
        - projects
      summary: 更新项目
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: 项目更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      tags:
        - projects
      summary: 删除项目
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: 项目删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Project deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    projectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    skip:
      name: skip
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    ProjectType:
      type: string
      enum:
        - research
        - analysis
        - report
        - documentation
        - custom
    ProjectStatus:
      type: string
      enum:
        - draft
        - planning
        - executing
        - paused
        - completed
        - archived
        - cancelled
...
    ProjectCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        type:
          $ref: '#/components/schemas/ProjectType'
        config:
          type: object
        objectives:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        config:
          type: object
        objectives:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        quality_gates:
          type: object
          additionalProperties:
            type: number
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/ProjectType'
        status:
          $ref: '#/components/schemas/ProjectStatus'
        config:
          type: object
        objectives:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        quality_gates:
          type: object
        progress:
          type: number
          minimum: 0
          maximum: 1
        quality_score:
          type: number
        success_rate:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
        status_code:
          type: integer
        error_code:
          type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...